working levels
i had to make the list of lists and had a bit of a struggle making the iteration work. initially with a  double for loop but then moving on to the single after realising it didnt work. had many issures with having the wrong self/rect/.y value as i was setting them but i managed to gett it to work. just need to add the level score in the game over screen